+++ src/engine_core/bullet_models/TerrainPhysics.h
@@ -13,7 +13,7 @@ private:
 
 public:
     btRigidBody* terrainRigidBody;
-    TerrainPhysics(int width, int length, unsigned short* heightData, btScalar minHeight, btScalar maxHeight);
+    TerrainPhysics(int width, int length, float* heightData, btScalar minHeight, btScalar maxHeight);
 };
 
 #endif
